Login

+++ TiendaComputadoras/routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\CategoriasController;
 use App\Http\Controllers\ColaboradoresController;
 use App\Http\Controllers\ColoresController;
 use App\Http\Controllers\Coloresproductos;
+use App\Http\Controllers\PaginaController;
 use App\Http\Controllers\PermisoController;
 use App\Http\Controllers\SalariosController;
 use App\Http\Controllers\ExportacionesController;
@@ -34,49 +35,53 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('welcome');
 });
+
+//Controller de validacion de inicio de session
+Route::get('/loginadmin',[PaginaController::class,'loginadmin'])->name('login');
+Route::post('/validarLogin', [PaginaController::class, 'validarLogin'])->name('validarLogin');
 //Controller del modulo de Gestion de Negocio
-Route::resource('cargos', CargosController::class)->parameters(['cargos' => 'cargos'])->names('cargos');
-Route::resource('colaboradores', ColaboradoresController::class)->parameters(['colaboradores' => 'colaboradores'])->names('colaboradores');
-Route::resource('asignaciones', AsignacionesController::class)->parameters(['asignaciones' => 'asignaciones'])->names('asignaciones');
-Route::resource('salarios', SalariosController::class)->parameters(['salarios' => 'salarios'])->names('salarios');
+Route::resource('cargos', CargosController::class)->parameters(['cargos' => 'cargos'])->names('cargos')->middleware('auth');
+Route::resource('colaboradores', ColaboradoresController::class)->parameters(['colaboradores' => 'colaboradores'])->names('colaboradores')->middleware('auth');
+Route::resource('asignaciones', AsignacionesController::class)->parameters(['asignaciones' => 'asignaciones'])->names('asignaciones')->middleware('auth');
+Route::resource('salarios', SalariosController::class)->parameters(['salarios' => 'salarios'])->names('salarios')->middleware('auth');
 //Controller de Catalogos
-Route::resource('categorias', CategoriasController::class)->parameters(['categorias' => 'categorias'])->names('categorias');
-Route::resource('subcategorias', SubcategoriasController::class)->parameters(['subcategorias' => 'subcategorias'])->names('subcategorias');
-Route::resource('marcas', MarcasController::class)->parameters(['marcas' => 'marcas'])->names('marcas');
-Route::resource('modelos', ModelosController::class)->parameters(['modelos' => 'modelos'])->names('modelos');
-Route::resource('colores', ColoresController::class)->parameters(['colores' => 'colores'])->names('colores');
-Route::resource('productos',ProductosController::class)->parameters(['productos' => 'productos'])->names('productos');
-Route::resource('coloresproductos',Coloresproductos::class)->parameters(['coloresproductos' => 'coloresproductos'])->names('coloresproductos');
-Route::get('/productos/{id}/multimedia', [ProductosController::class, 'multimedia'])->name('productos.multimedia');
-Route::post('/guardarmultimedia', [ProductosController::class, 'guardarmultimedia'])->name('productos.guardarmultimedia');
-Route::delete('/productos/destroyimg/{id}', [ProductosController::class, 'destroyimg'])->name('productos.destroyimg');
-Route::resource('precios',PreciosController::class)->parameters(['precios' => 'precios'])->names('precios');
+Route::resource('categorias', CategoriasController::class)->parameters(['categorias' => 'categorias'])->names('categorias')->middleware('auth');
+Route::resource('subcategorias', SubcategoriasController::class)->parameters(['subcategorias' => 'subcategorias'])->names('subcategorias')->middleware('auth');
+Route::resource('marcas', MarcasController::class)->parameters(['marcas' => 'marcas'])->names('marcas')->middleware('auth');
+Route::resource('modelos', ModelosController::class)->parameters(['modelos' => 'modelos'])->names('modelos')->middleware('auth');
+Route::resource('colores', ColoresController::class)->parameters(['colores' => 'colores'])->names('colores')->middleware('auth');
+Route::resource('productos',ProductosController::class)->parameters(['productos' => 'productos'])->names('productos')->middleware('auth');
+Route::resource('coloresproductos',Coloresproductos::class)->parameters(['coloresproductos' => 'coloresproductos'])->names('coloresproductos')->middleware('auth');
+Route::get('/productos/{id}/multimedia', [ProductosController::class, 'multimedia'])->name('productos.multimedia')->middleware('auth');
+Route::post('/guardarmultimedia', [ProductosController::class, 'guardarmultimedia'])->name('productos.guardarmultimedia')->middleware('auth');
+Route::delete('/productos/destroyimg/{id}', [ProductosController::class, 'destroyimg'])->name('productos.destroyimg')->middleware('auth');
+Route::resource('precios',PreciosController::class)->parameters(['precios' => 'precios'])->names('precios')->middleware('auth');
 
 
 //Controller del modulo de usuarios
-Route::resource('roles',RolesController::class)->parameters(['roles' => 'roles'])->names('roles');
-Route::resource('privilegios',PrivilegiosController::class)->parameters(['privilegios' => 'privilegios'])->names('privilegios');
-Route::resource('/permisos',PermisoController::class)->parameters(['permisos'=>'permisos'])->names('permisos');
-Route::resource('/usuarios',UsersController::class)->parameters(['usuarios'=>'usuarios'])->names('usuarios');
-Route::delete('/usuarios/destroyroles/{id}', [UsersController::class, 'destroyroles'])->name('usuarios.destroyroles');
+Route::resource('roles',RolesController::class)->parameters(['roles' => 'roles'])->names('roles')->middleware('auth');
+Route::resource('privilegios',PrivilegiosController::class)->parameters(['privilegios' => 'privilegios'])->names('privilegios')->middleware('auth');
+Route::resource('/permisos',PermisoController::class)->parameters(['permisos'=>'permisos'])->names('permisos')->middleware('auth');
+Route::resource('/usuarios',UsersController::class)->parameters(['usuarios'=>'usuarios'])->names('usuarios')->middleware('auth');
+Route::delete('/usuarios/destroyroles/{id}', [UsersController::class, 'destroyroles'])->name('usuarios.destroyroles')->middleware('auth');
 
 
 //Controller para multimedia
-Route::resource('img',ImgController::class)->parameters(['img' => 'img'])->names('img');
+Route::resource('img',ImgController::class)->parameters(['img' => 'img'])->names('img')->middleware('auth');
 //Rutas para exportacion y reportes
-Route::get('/exportcargosexcel', [ExportacionesController::class, 'exportcargosexcel'])->name('exportcargosexcel');
-Route::get('/exportcargopdf', [ExportacionesController::class, 'exportcargopdf'])->name('exportcargopdf');
-Route::get('/exportColaboradores', [ExportacionesController::class, 'exportColaboradores'])->name('exportColaboradores');
-Route::get('/exportColaboradorespdf', [ExportacionesController::class, 'exportColaboradorespdf'])->name('exportColaboradorespdf');
-Route::get('/exportasignaciones', [ExportacionesController::class, 'exportasignaciones'])->name('exportasignaciones');
-Route::get('/exportsalarios', [ExportacionesController::class, 'exportsalarios'])->name('exportsalarios');
-Route::get('/exportcategorias', [ExportacionesController::class, 'exportcategorias'])->name('exportcategorias');
-Route::get('/exportsubcategorias', [ExportacionesController::class, 'exportsubcategorias'])->name('exportsubcategorias');
-Route::get('/exportmarcas', [ExportacionesController::class, 'exportmarcas'])->name('exportmarcas');
-Route::get('/exportmodelos', [ExportacionesController::class, 'exportmodelos'])->name('exportmodelos');
-Route::get('/exportcolores', [ExportacionesController::class, 'exportcolores'])->name('exportcolores');
-Route::get('/exportproductos', [ExportacionesController::class, 'exportproductos'])->name('exportproductos');
-Route::get('/exportprecios', [ExportacionesController::class, 'exportprecios'])->name('exportprecios');
-Route::get('/exportaciones/pdf/{colaboradores}', [ExportacionesController::class, 'pdf'])->name('exportaciones.pdf');
-Route::get('/exportroles', [ExportacionesController::class, 'exportroles'])->name('exportroles');
-Route::get('/exportusuarios', [ExportacionesController::class, 'exportusuarios'])->name('exportusuarios');
+Route::get('/exportcargosexcel', [ExportacionesController::class, 'exportcargosexcel'])->name('exportcargosexcel')->middleware('auth');
+Route::get('/exportcargopdf', [ExportacionesController::class, 'exportcargopdf'])->name('exportcargopdf')->middleware('auth');
+Route::get('/exportColaboradores', [ExportacionesController::class, 'exportColaboradores'])->name('exportColaboradores')->middleware('auth');
+Route::get('/exportColaboradorespdf', [ExportacionesController::class, 'exportColaboradorespdf'])->name('exportColaboradorespdf')->middleware('auth');
+Route::get('/exportasignaciones', [ExportacionesController::class, 'exportasignaciones'])->name('exportasignaciones')->middleware('auth');
+Route::get('/exportsalarios', [ExportacionesController::class, 'exportsalarios'])->name('exportsalarios')->middleware('auth');
+Route::get('/exportcategorias', [ExportacionesController::class, 'exportcategorias'])->name('exportcategorias')->middleware('auth');
+Route::get('/exportsubcategorias', [ExportacionesController::class, 'exportsubcategorias'])->name('exportsubcategorias')->middleware('auth');
+Route::get('/exportmarcas', [ExportacionesController::class, 'exportmarcas'])->name('exportmarcas')->middleware('auth');
+Route::get('/exportmodelos', [ExportacionesController::class, 'exportmodelos'])->name('exportmodelos')->middleware('auth');
+Route::get('/exportcolores', [ExportacionesController::class, 'exportcolores'])->name('exportcolores')->middleware('auth');
+Route::get('/exportproductos', [ExportacionesController::class, 'exportproductos'])->name('exportproductos')->middleware('auth');
+Route::get('/exportprecios', [ExportacionesController::class, 'exportprecios'])->name('exportprecios')->middleware('auth');
+Route::get('/exportaciones/pdf/{colaboradores}', [ExportacionesController::class, 'pdf'])->name('exportaciones.pdf')->middleware('auth');
+Route::get('/exportroles', [ExportacionesController::class, 'exportroles'])->name('exportroles')->middleware('auth');
+Route::get('/exportusuarios', [ExportacionesController::class, 'exportusuarios'])->name('exportusuarios')->middleware('auth');
